added client side search options

+++ LifelineAssistant/main/AdminRoutes.py
@@ -4,11 +4,11 @@ from . import main
 from .Forms import MessageForm
 from datetime import datetime
 
+#mysql = MySQLdb.connect(host='35.231.239.49', user='root', passwd='tGHC97h8xDoI6b1m', db='assurance-wireless-db')
 mysql = MySQLdb.connect(host='localhost', user='root', passwd='mysql', db='awla_db')
-
-# -------------------------------------------------------------------------+
-# Application Management Section                                           |
-# -------------------------------------------------------------------------+
+# ---------------------------------------------------------------------------------+
+# Application Management Section                                                   |
+# ---------------------------------------------------------------------------------+
 
 @main.route("/applications")
 def show_apps():
@@ -29,8 +29,6 @@ def show_apps():
 def search_apps():
     """search by application status"""
     if session['admin'] and request.method == "POST":
-        if request.form.get('filter') == 'All':
-            return show_apps()
         curs = mysql.cursor()
         sql = """select appid, status, fname, lname, 
         applicant_email, zipcode, created from applications
@@ -47,10 +45,10 @@ def search_apps():
 def app_comments(app, reason):
     """provide reason for accepting or denying applications"""
     curs = mysql.cursor()
-    sql = "insert into reports(body, admin_email, applicant_email) values (%s, %s, %s)"
+    sql = "insert into reports(reasoning, admin_email, applicant_email) values (%s, %s, %s)"
     values = [reason, session.get('email'), app]
     curs.execute(sql, values)
-    mysql.commit()
+    mysql.commit()  
 
 
 @main.route("/process_apps", methods=['GET', 'POST'])
@@ -58,8 +56,8 @@ def process_apps():
     """ modify application data """
     if session['admin'] and request.method == "POST":
         # search by status: accepted, denied, status
-        if "search-btn" in request.form:
-            if request.form.get('filter') == 'All':
+        if "filter" in request.form:
+            if request.form.get('filter') == 'Show All':
                 return show_apps()
             else:
                 return search_apps() 
@@ -79,9 +77,41 @@ def process_apps():
             return show_apps()  # get results from modal then redirect view
         return render_template("admin_templates/applications.html")
 
-# -------------------------------------------------------------------------+
-# Customer Support Respondent Section                                       |
-# -------------------------------------------------------------------------+
+
+# ----------------------------------------------------------------------------------+    
+# Customer Support Ticket Responder Section                                         |
+# ----------------------------------------------------------------------------------+
+
+def filter_status():
+    """ Filter by Ticket Status, function is similar to search_apps """
+    if session['admin'] and request.method == 'POST':
+        if 'tkt' in request.form:
+            return show_tickets()
+        curs = mysql.cursor()
+        sql = """select id, status, category, question, requester, acceptor, time_created 
+            from support_tickets where status = %s order by time_created asc"""
+        values = [request.form.get('filter')]
+        
+        curs.execute(sql, values)
+        mysql.commit()
+        tickets = []
+        for row in curs:
+            tickets.append(row)
+        return tickets
+
+
+def filter_categories():
+    curs = mysql.cursor()
+    sql = """select id, status, category, question, requester, acceptor, time_created 
+            from support_tickets where category = %s order by time_created asc"""
+    values = [request.form.get('filter2')]
+    curs.execute(sql, values)
+    mysql.commit()
+    tickets = []
+    for row in curs:
+        tickets.append(row)
+    return tickets
+
 
 @main.route('/comment-queue', methods=['POST', 'GET'])
 def show_tickets():
@@ -94,17 +124,16 @@ def show_tickets():
         for row in curs:
             tickets.append(row)
         mysql.commit()
-        if request.method == 'POST':
-        # if name=dropdown option, modify status
-            if 'mod-status' in request.form:
-                sql = "update table support_tickets set status = %s where id = %s"
-                values = [request.form.get('mod-status')]
-                print(request.form.get('mod-status'))
-                print(request.form['mod-status'])
-        # if name=tkt return ticket responder view
-            if 'tkt' in request.form:
-                ticket_id = request.form.get('tkt')
-                return redirect(url_for('.select_ticket', ticket_id=ticket_id))
+
+        if 'filter' in request.form and request.form['filter'] != 'Show All':
+            tickets = filter_status()
+        if 'filter2' in request.form:
+            tickets = filter_categories()
+
+        # move to ticket detail view if ticket is clicked
+        if 'tkt' in request.form:
+            ticket_id = request.form.get('tkt')
+            return redirect(url_for('.select_ticket', ticket_id=ticket_id))
         return render_template('admin_templates/comment-queue.html', tickets=tickets)
 
 
@@ -137,9 +166,10 @@ def select_ticket(ticket_id):
         tkt_info = get_ticket_info(ticket_id)
         return render_template('admin_templates/ticket-response.html', tkt_info=tkt_info)
 
-# -------------------------------------------------------------------------+
-# Customer Support Respondent Section, Real Time Chat                      |
-# -------------------------------------------------------------------------+
+
+# ----------------------------------------------------------------------------------+
+# Customer Support Respondent Section, Real Time Chat                               |
+# ----------------------------------------------------------------------------------+
 
 def format_html(data):
     html = ""
@@ -157,7 +187,7 @@ def format_html(data):
                 if count == 0:
                     html += '<div>'
                 count += 1
-                html += "<p>" + x + "</p>"
+                html += '<p>' + x + '</p>'
                 if count == 2:
                     html += '</div>'
         count = 0
@@ -193,20 +223,24 @@ def show_data():
     data = format_html(msg_data)
     return data
 
-# -------------------------------------------------------------------------+
-# Support Ticket Actions: accept ticket, change ticket status, remove      |
-# -------------------------------------------------------------------------+
+
+# ----------------------------------------------------------------------------------+
+# Support Ticket Actions: accept ticket, change ticket status, remove               |
+# ----------------------------------------------------------------------------------+
 
 @main.route('/message-action', methods=['POST'])
 def message_action():
-    if request.method == 'POST':
+    if request.method == 'POST' and session['admin']:
         curs = mysql.cursor()
         ticket_id = session.get('tkt')
-        if 'take-ticket' in request.form:
+
+        if 'mod-status' in request.form:
+            print(request.form['mod-status'])
             sql = "update support_tickets set status = %s, acceptor = %s where id = %s"
-            values = ['in progress', session.get('email'), ticket_id]
+            values = [request.form['mod-status'], session.get('email'), ticket_id]
             curs.execute(sql, values)
             mysql.commit()
+
         if 'delete-ticket' in request.form:
             sql = "delete from support_messages where ticket_id = %s"
             curs.execute(sql, [ticket_id])
